feature: DCMAW-21004 expand test coverage (#866)
This code change merely expands the test coverage of the app
	to demonstrate the scenario is an accepted behaviour.

	Specifically:
		When an a `SecureStoreError(.cantRetrieveKey)` error is thrown
		as part of an attempt to start a session,
		that error will indeed lead to failure to start the session
		and an analytic will be logged.
