+++ Tests/UnitTests/Login/WebAuthenticationServiceTests.swift
@@ -240,6 +240,38 @@ struct WebAuthenticationServiceTests {
         #expect(mockAnalyticsService.crashesLogged.count == 1)
     }
     
+    /// GIVEN I am "a returning user", who is "NOT enrolling" and "NOT authenticated" due to a `nil` `expiryDate`
+    /// AND `SecureStoreService` throws `SecureStoreError(.cantRetrieveKey)` when `saveItem` is called as part of `PersistentSessionManager.saveAuthSession()`
+    /// WHEN I start a web session
+    /// THEN an error is logged as a crash in analytics
+    /// AND the error is thrown
+    @Test
+    func test_startWebSession_throws_cannotRetrieveKey() async throws {
+        let mockAnalyticsService = MockAnalyticsService()
+
+        let mockEncryptedStore = MockSecureStoreService()
+        
+        let sessionManager: PersistentSessionManager = try .makeNonReturningNonEnrollingUnauthenticatedUserWithoutSavedExpiryDate(
+            mockAnalyticsService: mockAnalyticsService,
+            mockEncryptedStore: mockEncryptedStore
+        )
+        
+        mockEncryptedStore.errorFromSaveItem = SecureStoreError(.cantRetrieveKey)
+        
+        let sut: WebAuthenticationService = await .make(
+            sessionManager: sessionManager,
+            mockAnalyticsService: mockAnalyticsService
+        )
+        
+        let error = await #expect(throws: SecureStoreError.self) {
+            try await sut.startWebSession(appIntegrityProvider: AppIntegrityProviderStub())
+        }
+        
+        #expect(error?.kind == .cantRetrieveKey)
+        
+        #expect(mockAnalyticsService.crashesLogged.count == 1)
+    }
+    
     @Test
     func test_startWebSession_success() async throws {
         let sessionManager: PersistentSessionManager = .make()
@@ -307,83 +339,5 @@ extension WebAuthenticationService {
         mockLoginSession.errorFromFinalise = error
         
         return .make(window: window, mockLoginSession: mockLoginSession)
-    }
-}
-
-extension PersistentSessionManager {
-    /// Creates a `PersistentSessionManager` with the following conditions:
-    /// * `persistentID = nil` i.e. not stored in the `encryptedStore`
-    /// * `isReturningUser = true`
-    /// *  `walletSDK.isEmpty = true`
-    ///
-    /// A call to `startAuthSession(:using:)` assumes this is "a returning user" that cannot authenticate due to
-    /// the missing `persistendId` results  in call to `clearAllSessionData` to delete my session & Wallet data.
-    static func makeReturningUnauthenticatedUser(mockAnalyticsService: MockAnalyticsService = MockAnalyticsService(),
-                                                 mockEncryptedStore: MockSecureStoreService = MockSecureStoreService(),
-                                                 mockUnprotectedStore: (any DefaultsStoring & SessionBoundData) = MockDefaultsStore(),
-                                                 walletSessionData: SessionBoundData = WalletSessionBoundDataStub(),
-                                                 refreshTokenExchangeManager: TokenExchangeManaging = MockRefreshTokenExchangeManager(),
-                                                 analyticsPreferenceStore: (any AnalyticsPreferenceStore & SessionBoundData) = MockAnalyticsPreferenceStore()
-    ) throws -> PersistentSessionManager {
-        mockUnprotectedStore.set(
-            true,
-            forKey: OLString.returningUser
-        )
-        
-        let walletSDK = MockWalletSDKWrapper()
-        walletSDK.isEmpty = true
-        
-        return try .make(encryptedStore: mockEncryptedStore,
-                         unprotectedStore: mockUnprotectedStore,
-                         analyticsService: mockAnalyticsService,
-                         walletSDK: walletSDK,
-                         walletSessionData: walletSessionData,
-                         refreshTokenExchangeManager: refreshTokenExchangeManager,
-                         serialTaskQueue: SerialTaskQueue(),
-                         analyticsPreferenceStore: analyticsPreferenceStore)
-    }
-    
-    /// Creates a `PersistentSessionManager` with the following conditions:
-    /// * `isEnrolling = false`
-    /// * `isReturningUser = true`
-    /// * `persistentID = [random uuid]`
-    /// *  `walletSDK.isEmpty = true`
-    ///
-    /// A call to `startAuthSession(:using:)` assumes this is "a returning user" with a `sessionState` that is `.nonePresent` due to a missing `expiryDate`.
-    static func makeNonReturningNonEnrollingUnauthenticatedUserWithoutSavedExpiryDate(
-        mockAnalyticsService: MockAnalyticsService = MockAnalyticsService(),
-        accessControlEncryptedSecureStoreMigrator: MockSecureStoreService = MockSecureStoreService(),
-        mockEncryptedStore: MockSecureStoreService = MockSecureStoreService(),
-        mockUnprotectedStore: (any DefaultsStoring & SessionBoundData) = MockDefaultsStore(),
-        walletSessionData: SessionBoundData = WalletSessionBoundDataStub(),
-        refreshTokenExchangeManager: TokenExchangeManaging = MockRefreshTokenExchangeManager(),
-        analyticsPreferenceStore: (any AnalyticsPreferenceStore & SessionBoundData) = MockAnalyticsPreferenceStore()
-    ) throws -> PersistentSessionManager {
-        mockUnprotectedStore.set(
-            true,
-            forKey: OLString.returningUser
-        )
-
-        try mockEncryptedStore.saveItem(
-            item: UUID().uuidString,
-            itemName: OLString.persistentSessionID
-        )
-        
-        let walletSDK = MockWalletSDKWrapper()
-        walletSDK.isEmpty = true
-        
-        let persistentSessionManager: PersistentSessionManager = try .make(
-                         accessControlEncryptedSecureStoreMigrator: accessControlEncryptedSecureStoreMigrator,
-                         encryptedStore: mockEncryptedStore,
-                         unprotectedStore: mockUnprotectedStore,
-                         analyticsService: mockAnalyticsService,
-                         walletSDK: walletSDK,
-                         walletSessionData: walletSessionData,
-                         refreshTokenExchangeManager: refreshTokenExchangeManager,
-                         serialTaskQueue: SerialTaskQueue(),
-                         analyticsPreferenceStore: analyticsPreferenceStore)
-        
-        persistentSessionManager.isEnrolling = false
-        return persistentSessionManager
     }
 }
